Refactor search router into service
