+++ tests/search/vespa/test_whole_database_download.py
@@ -127,9 +127,7 @@ def test_all_data_download_fails_when_s3_upload_failed(
         patch("app.api.api_v1.routers.search.PIPELINE_BUCKET", "test_pipeline_bucket"),
         patch("app.api.api_v1.routers.search.DOCUMENT_CACHE_BUCKET", "test_cdn_bucket"),
         patch("app.clients.aws.client.S3Client.is_connected", return_value=True),
-        patch(
-            "app.api.api_v1.routers.search._get_s3_doc_url_from_cdn", return_value=None
-        ),
+        patch("app.service.search.get_s3_doc_url_from_cdn", return_value=None),
     ):
         data_client.follow_redirects = False
         download_response = data_client.get(ALL_DATA_DOWNLOAD_ENDPOINT, headers=headers)
